+++ src/app/cart/guards/is-cart-empty.guard.ts
@@ -1,9 +1,10 @@
 import { inject } from '@angular/core';
 import { CanActivateFn } from '@angular/router';
-import { CartService } from '../services/cart.service';
+import { map } from 'rxjs';
+import { CartObservableService } from '../services/cart-observable.service';
 
 export const IsCartEmptyGuard: CanActivateFn = (route, state) => {
-  const cartService = inject(CartService);
-  const isEmpty = cartService.isEmptyCart();
-  return isEmpty;
+  const cartService = inject(CartObservableService);
+  let isEmpty = false;
+  return cartService.isEmptyCart().pipe(map(empty => (isEmpty = !empty)));
 };