+++ src/app/core/services/local-storage.service.ts
@@ -4,8 +4,12 @@ import { Injectable } from '@angular/core';
   providedIn: 'root'
 })
 export class LocalStorageService {
-  setItem(key: string, value: string) {
+  setItem(key: string, value: any) {
     localStorage.setItem(key, value);
+  }
+
+  getItem(key: string): string | null {
+    return localStorage.getItem(key);
   }
 
   removeItem(key: string) {