+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,10 +1,8 @@
 import { Component, OnInit, inject } from '@angular/core';
-import { Observable, of } from 'rxjs';
-import { ProductsService } from '../../services/products.service';
 import { Product } from '../../models/product.model';
-import { CartService } from '../../../cart/services/cart.service';
 import { Router } from '@angular/router';
-import { CommonModule } from '@angular/common';
+import { ProductsPromiseService } from '../../services/products-promise.service';
+import { CartObservableService } from 'src/app/cart/services/cart-observable.service';
 
 @Component({
   selector: 'app-product-list',
@@ -12,10 +10,10 @@ import { CommonModule } from '@angular/common';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-  products$: Observable<Array<Product>> = of([]);
+  products$!: Promise<Product[]>;
 
-  private productsService = inject(ProductsService);
-  private cartService = inject(CartService);
+  private productsService = inject(ProductsPromiseService);
+  private cartService = inject(CartObservableService);
   private router = inject(Router);
 
   ngOnInit(): void {
@@ -24,7 +22,7 @@ export class ProductListComponent implements OnInit {
 
   onAddToCart(product: Product) {
     alert(product.name + ' added to the cart');
-    this.cartService.addCartItem(product);
+    this.cartService.addCartItem(product).subscribe();
   }
 
   onViewProduct(product: Product) {